fix sql and git token

+++ dss-orchestrator/orchestrators/dss-workflow/dss-workflow-server/src/main/java/com/webank/wedatasphere/dss/workflow/service/impl/DSSFlowServiceImpl.java
@@ -475,7 +475,9 @@ public class DSSFlowServiceImpl implements DSSFlowService {
 
         }
         // 先删后增
-        nodeContentUIMapper.deleteNodeContentUIByContentList(contentIdListByOrchestratorId);
+        if (CollectionUtils.isNotEmpty(contentIdListByOrchestratorId)) {
+            nodeContentUIMapper.deleteNodeContentUIByContentList(contentIdListByOrchestratorId);
+        }
         nodeContentUIMapper.batchInsertNodeContentUI(nodeContentUIDOS);
     }
